+++ buildSrc/src/main/java/com/ivy/wallet/buildsrc/Project.kt
@@ -1,0 +1,16 @@
+package com.ivy.wallet.buildsrc
+
+
+object Project {
+    //Version
+    const val versionName = "4.4.4"
+    const val versionCode = 144
+
+    //Compile SDK & Build Tools
+    const val compileSdkVersion = 34
+
+    //App
+    const val applicationId = "com.ivy.wallet"
+    const val minSdk = 28
+    const val targetSdk = 34
+}
